lyf commits 教师创建课程，修改课程，删除课程 创建讨论课，修改讨论课，删除讨论课 创建班级，修改班级，删除班级 删去了teacher student的updatePassword的方法 还没测试

+++ src/main/java/seminar/dao/AdminDAO.java
@@ -23,26 +23,43 @@ public class AdminDAO {
         return adminMapper.selectAdministratorByAdminName(name);
     }
 
-
     /**
      * @author lyf
      */
-    public void addAdmin(Administrator administrator){
-        adminMapper.insertAdministrator(administrator);
+    public boolean create(Administrator administrator, String adminName){
+
+        if(adminMapper.selectAdministratorByAdminName(adminName).isEmpty()) {
+            adminMapper.insertAdministrator(administrator);
+            return true;
+        }
+        else return false;
     }
 
     /**
      * @author lyf
      */
-    public void deleteAdminByName(String adminName){
-        adminMapper.deleteAdministratorByAdminName(adminName);
+    public boolean deleteByName(String adminName){
+        if(adminMapper.selectAdministratorByAdminName(adminName).isEmpty())
+            return false;
+        else{
+            adminMapper.deleteAdministratorByAdminName(adminName);
+            return true;
+        }
+
     }
 
     /**
      * @author lyf
      */
-    public void updateAdmin(Administrator administrator){
-        adminMapper.updateAdministrator(administrator);
+    public boolean update(Administrator administrator, String adminName){
+        if(adminMapper.selectAdministratorByAdminName(adminName).isEmpty())
+            return false;
+
+        else{
+            adminMapper.updateAdministrator(administrator);
+            return true;
+        }
+
     }
 
 }
