+++ src/main/java/seminar/dao/ClbumDao.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import seminar.entity.Clbum;
 import seminar.mapper.ClbumMapper;
+import seminar.entity.Course;
+import seminar.mapper.CourseMapper;
 
 import java.util.List;
 
@@ -13,42 +15,69 @@ import java.util.List;
 @Component
 public class ClbumDao {
     private final ClbumMapper clbumMapper;
+    private final CourseMapper courseMapper;
 
     @Autowired
-    public ClbumDao(ClbumMapper clbumMapper) {
+    public ClbumDao(ClbumMapper clbumMapper, CourseMapper courseMapper) {
         this.clbumMapper = clbumMapper;
+        this.courseMapper = courseMapper;
     }
 
-    public List<Clbum> getClbumsByCourseId(String courseId){
+    public List<Clbum> getByCourseId(String courseId){
         return clbumMapper.selectClbumByCourseId(courseId);
     }
 
     /**
      * @author lyf
      */
-    public void createClbum(Clbum clbum){
+    public boolean create(Clbum clbum){
+        List<Clbum> clbums = clbumMapper.selectClbumByCourseId(clbum.getCourseId());
+        for(Clbum c:clbums){
+            if(c.getClbumName().equals(clbum.getClbumName()))
+                return false;
+        }
+
         clbumMapper.insertClbum(clbum);
+        return true;
     }
 
     /**
      * @author lyf
      */
-    public void updateClbum(Clbum clbum){
+    public boolean update(Clbum clbum){
+        List<Clbum> clbums = clbumMapper.selectClbumByCourseId(clbum.getCourseId());
+        for(Clbum c:clbums){
+            if(c.getClbumName().equals(clbum.getClbumName()))
+                return false;
+        }
         clbumMapper.updateClbum(clbum);
+        return true;
     }
 
     /**
      * @author lyf
      */
-    public void deleteClbumById(String clbumId){
-        clbumMapper.deleteClbumById(clbumId);
+    public boolean deleteById(String clbumId){
+        List<Clbum> clbums = clbumMapper.selectClbumById(clbumId);
+        if(clbums.isEmpty())
+            return false;
+        else {
+            clbumMapper.deleteClbumById(clbumId);
+            return true;
+        }
     }
 
     /**
      * @author lyf
      */
-    public void deleteClbumByCourseId(String courseId){
-        clbumMapper.deleteClbumByCourseId(courseId);
+    public boolean deleteByCourseId(String courseId){
+        List<Course> courses = courseMapper.selectCourseById(courseId);
+        if(courses.isEmpty())
+            return false;
+        else{
+            clbumMapper.deleteClbumByCourseId(courseId);
+            return true;
+        }
     }
 
 }
