+++ src/main/java/seminar/dao/CourseDAO.java
@@ -3,10 +3,9 @@ package seminar.dao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import seminar.entity.Course;
-import seminar.entity.MaxMinRegulation;
 import seminar.mapper.CourseMapper;
-import seminar.mapper.MaxMinRegulationMapper;
 
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -21,29 +20,53 @@ public class CourseDAO {
         this.courseMapper = courseMapper;
     }
 
-    public List<Course> getCoursesByTeacherId(String teacherId) {
+    public List<Course> getByTeacherId(String teacherId) {
         return courseMapper.selectCourseByTeacherId(teacherId);
     }
-
-
+    /**
+    * @author lyf
+    */
+    public boolean deleteByTeacherId(String teacherId){
+        List<Course> courses = courseMapper.selectCourseByTeacherId(teacherId);
+        if(courses.isEmpty()) return false;
+        else{
+            courseMapper.deleteCourseByTeacherId(teacherId);
+            return true;
+        }
+    }
     /**
      * @author lyf
      */
-    public void createCourse(Course course) {
+    public boolean create(Course course) {
+        List<Course> courses = courseMapper.selectCourseByTeacherId(course.getTeacherId());
+        for(Course c: courses) {
+            if (c.getCourseName().equals(course.getCourseName()))
+                return false;
+        }
         courseMapper.insertCourse(course);
+        return true;
     }
 
     /**
      * @author lyf
      */
-    public void deleteCourseById(String courseId){
-        courseMapper.deleteCourseById(courseId);
+    public boolean deleteByCoruseId(String courseId){
+        if(courseMapper.selectCourseById(courseId).isEmpty()) {
+            courseMapper.deleteCourseById(courseId);
+            return true;
+        }
+        else return false;
     }
 
     /**
      * @author lyf
      */
-    public void updateCourseById(Course course){
-        courseMapper.updateCourse(course);
+    public boolean update(Course course) {
+        if (courseMapper.selectCourseById(course.getId()).isEmpty()) {
+            courseMapper.updateCourse(course);
+            return true;
+        }
+        else return false;
     }
+
 }
